Update: replace all os.path.normcase with os.path.normpath
This will fix filenames being forced into lowercase on Windows. From the
python docs:

normcase: `On Windows, convert all characters in the pathname to
lowercase, and also convert forward slashes to backward slashes. On
other operating systems, return the path unchanged.`

normpath: `Normalize a pathname by collapsing redundant separators and
up-level references so that A//B, A/B/, A/./B and A/foo/../B all become
A/B. This string manipulation may change the meaning of a path that
contains symbolic links. On Windows, it converts forward slashes to
backward slashes.`